% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_deaminations.R
\name{annotate_deaminations}
\alias{annotate_deaminations}
\title{Annotate deaminations}
\usage{
annotate_deaminations(
  predictions,
  format = "tsv",
  outfolder = ".",
  outname = "ideafix_labels",
  vcf_filename = NULL
)
}
\arguments{
\item{predictions}{predictions generated by filter_variants}

\item{format}{exported file format. Can be "tsv" or "vcf". Defaults to "tsv".}

\item{outfolder}{folder to output the exported file. Defaults to current working directory}

\item{outname}{filename of exported file. Defaults to "ideafix_labels.txt"}

\item{vcf_filename}{if format is "vcf", path to the vcf file ideafix has been run over}
}
\value{
None
}
\description{
Exports prediction results to a text file
}
\details{
In case vcf is selected, the vcf must be provided the results will be appended to input vcf file...
}
