% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_append.R
\name{append_vcf}
\alias{append_vcf}
\title{Title}
\usage{
append_vcf(vcf_filename, predictions)
}
\arguments{
\item{vcf_filename}{input vcf}

\item{predictions}{predictions generated by filter_variants}
}
\value{

}
\description{
Title
}
