% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{check_algorithm}
\alias{check_algorithm}
\title{check_algorithm}
\usage{
check_algorithm(algorithm)
}
\arguments{
\item{algorithm}{One of the algorithms to use to filter the variants. Can be RF or XGBoost.}
}
\value{
None
}
\description{
check_algorithm
}
