% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_arguments}
\alias{check_arguments}
\title{Check required arguments exist}
\usage{
check_arguments(defined, passed)
}
\arguments{
\item{defined}{Arguments required for the function to run}

\item{passed}{Arguments passed to the function}
}
\value{
None
}
\description{
Checks if required arguments for a function are present
}
\details{
It stops the program if the argument is not present.
}
