% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{check_descriptors}
\alias{check_descriptors}
\title{check_descriptors}
\usage{
check_descriptors(variant_descriptors, algorithm)
}
\arguments{
\item{variant_descriptors}{Matrix containing the descriptors of the variants to be analyzed and subjected to filtering}

\item{algorithm}{One of the algorithms to use to filter the variants. Can be RF or XGBoost.}
}
\value{
None
}
\description{
check_descriptors
}
