% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{check_lowercase}
\alias{check_lowercase}
\title{Check case}
\usage{
check_lowercase(char)
}
\arguments{
\item{char}{character string}
}
\value{
Logical value indicating whether \strong{char} is lowercase.
}
\description{
\code{check_lowercase} checks whether a character is lowercase.
}
