% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_variants.R
\name{classify_variants_RF}
\alias{classify_variants_RF}
\title{Classify variants as deaminations or non-deaminations with Random Forest}
\usage{
classify_variants_RF(variant_descriptors)
}
\arguments{
\item{variant_descriptors}{tibble containing the variants to be classified together with their values for a series of descriptors obtained by \code{\link{get_descriptors}}.}
}
\value{
Tibble with six columns: CHROM, POS, REF, ALT, DEAM_SCORE, DEAMINATION. CHROM and POS identify the variant position, REF and ALT describe the reference and alternate alleles. DEAM_SCORE equals to the deamination score yielded by the selected classification algorithm (RF or XGBoost). Note that these values should not be interpreted as ordinary probabilities. DEAMINATION contains the label ideafix has assigned to the variant based on an optimized classification threshold.
}
\description{
\code{classify_variants_RF} classifies a set of C:G > T:A variants into deaminations or non-deaminations based on a series of relevant variant descriptors, using a Random Forest-based model.
}
\details{
As opposed to classify_variants_XGBoost, \code{classify_variants_RF} can run with a \code{variant_descriptors} object with fewer descriptors than those included in the model. The whole set of descriptors includes VAF, number of alternate bases, normalized number of alternate bases, number of reference bases, normalized number of reference bases, reference allele, alternate allele, base quality, base quality fraction, fragment length, median position from read end, mapping quality, FDeamC, SOB, SB-GUO, SB-GATK, normalized median position from read end, base two positions before, base one position before, base two positions after, base one position after, dinucleotide before and dinucleotide after. However, if that is the case, good performance is not guaranteed.
}
