% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{filter_variants}
\alias{filter_variants}
\title{filter_variants}
\usage{
filter_variants(variant_descriptors, algorithm)
}
\arguments{
\item{variant_descriptors}{Matrix containing the descriptors of the variants to be analyzed and subjected to filtering}

\item{algorithm}{One of the algorithms to use to filter the variants. Can be RF or XGBoost.}
}
\value{
Matrix containing three columns: variant id, probability of deamination and predicted label: deamination or non-deamination
}
\description{
filter_variants
}
