% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{filter_variants_RF}
\alias{filter_variants_RF}
\title{filter_variants_RF}
\usage{
filter_variants_RF(variant_descriptors)
}
\arguments{
\item{variant_descriptors}{Matrix containing the descriptors of the variants to be analyzed and subjected to filtering}
}
\value{
Matrix containing three columns: variant id, probability of deamination and predicted label: deamination or non-deamination
}
\description{
filter_variants_RF
}
\details{
As opposed to filter_variants_XGBoost, filter_variants_RF can run even if not all the descriptors in the model are included in \strong{variant_descriptors}. However, good performance is not guaranteed.
}
