% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{filter_variants_XGBoost}
\alias{filter_variants_XGBoost}
\title{filter_variants_XGBoost}
\usage{
filter_variants_XGBoost(variant_descriptors)
}
\arguments{
\item{variant_descriptors}{Matrix containing the descriptors of the variants to be analyzed and subjected to filtering}
}
\value{
Matrix containing three columns: variant id, probability of deamination and predicted label: deamination or non-deamination
}
\description{
filter_variants_XGBoost
}
\details{
Note that the function only works if \strong{variant_descriptors} includes all the descriptors that were used to train the model
}
