% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_AF}
\alias{get_AF}
\title{Get allele frequency}
\usage{
get_AF(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}

\item{...}{Not to be checked}
}
\value{
Vector containing the allele frequency (AF) of each variant, ordered in order of appearance in the vcf file
}
\description{
Get allele frequency
}
