% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_alt_allele}
\alias{get_alt_allele}
\title{Get alternate allele}
\usage{
get_alt_allele(vcf_filename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{...}{Not to be checked}
}
\value{
Vector containing the alternate allele of each variant, ordered in order of appearance in the vcf file
}
\description{
Get alternate allele
}
