% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_alt_bases}
\alias{get_alt_bases}
\title{Get number of alternate alleles}
\usage{
get_alt_bases(vcf_filename, depth, samplename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{depth}{Vector containing the read depth at each variant locus}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}

\item{...}{Not to be checked}
}
\value{
List containing two vectors: the number of reads that support the alternate allele and ...
}
\description{
Get number of reads that support the alternate allele
}
