% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_depth}
\alias{get_depth}
\title{Get depth}
\usage{
get_depth(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}

\item{...}{Not to be checked}
}
\value{
Vector containing the read depth at each variant locus
}
\description{
Gets read depth at variant locus
}
