% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_fraglength}
\alias{get_fraglength}
\title{Get variant median fragment length}
\usage{
get_fraglength(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}

\item{...}{Not to be checked}
}
\value{
Vector containing the median median fragment length of each variant, ordered in order of appearance in the vcf file
}
\description{
Get median fragment length for reads that support the alternate allele
}
