% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_map_qual}
\alias{get_map_qual}
\title{Retrieve mapping quality}
\usage{
get_map_qual(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{character string naming the path to the input vcf, i.e. the vcf file containing the variants to classify. This file must have been generated with Mutect2, either in tumor only or tumor/normal mode with strand bias annotation enabled.}

\item{samplename}{character string naming the sample in \code{vcf_filename}. This must match the name given to the sample when running Mutect2. It can be obtained by calling \code{get_samplename} on \code{vcf_filename}.}

\item{...}{additional arguments not to be checked.}
}
\value{
Numeric vector containing the MAPQ mapping quality of each C:G > T:A variant in \code{vcf_filename}, ordered by appearance in the file.
}
\description{
\code{get_map_qual} retrieves, for each C:G > T:A variant present in \code{vcf_filename}, the median Phred mapping quality (MAPQ) of the bases that support the alternate allele. This data is retrieved from \code{vcf_filename}.
}
