% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_ref_allele}
\alias{get_ref_allele}
\title{Get reference allele}
\usage{
get_ref_allele(vcf_filename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{...}{Not to be checked}
}
\value{
Vector containing the reference allele of each variant, ordered in order of appearance in the vcf file
}
\description{
Get reference allele
}
