% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_RPOB}
\alias{get_RPOB}
\title{Calculate read pair orientation bias}
\usage{
get_RPOB(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{character string naming the path to the input vcf, i.e. the vcf file containing the variants to classify. This file must have been generated with Mutect2, either in tumor only or tumor/normal mode with strand bias annotation enabled.}

\item{samplename}{character string naming the sample in \code{vcf_filename}. This must match the name given to the sample when running Mutect2. It can be obtained by calling \code{get_samplename} on \code{vcf_filename}.}

\item{...}{additional arguments not to be checked.}
}
\value{
List containing two numeric vectors: FDeamC and SOB. Each element in the vectors corresponds to a C:G > T:A variant in \code{vcf_filename} and the elements are ordered by appearance in \code{vcf_filename}.
}
\description{
\code{get_RPOB} calculates, for each C:G > T:A variant present in \code{vcf_filename}, two read pair orientation bias (RPOB) metrics: FdeamC and strand orientation bias (SOB). The calculation is based on data retrieved from \code{vcf_filename}.
}
