% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_surr_bases}
\alias{get_surr_bases}
\title{Get surrounding bases}
\usage{
get_surr_bases(fasta_filename, mutation_ids, k)
}
\arguments{
\item{fasta_filename}{Reference genome data was aligned to.}

\item{mutation_ids}{Vector that contains the mutation ids of the variants to be subjected to filtering. Mutation id format is samplename:chr:pos.}

\item{k}{Number of bases to the right and to the left of the variant locus to get the genomic sequence from.}
}
\value{
Genomic bases \strong{k} positions before and after each variant in \strong{mutation_ids}, including the variant locus itself.
}
\description{
Returns the base composition of the fragment that surrounds each of the provided genomic variants. The region goes from the locus - \strong{k} to locus + \strong{k}, including the locus itself.
}
