% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideafix.R
\docType{package}
\name{ideafix}
\alias{ideafix}
\title{ideafix, DEAmination FIXing Package}
\description{
ideafix is a decision tree-based variant refinement tool that filters
formaldehyde-induced cytosine deaminations from variant lists obtained from
DNA sequencing data from FFPE specimens
}
\author{
Maitena Tellaetxe Abete \email{maitenat@gmail.com}

Borja Calvo Molinos \email{borja.calvo@ehu.eus}
}
