% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_variants.R
\name{join_uncompress_targz}
\alias{join_uncompress_targz}
\title{Join and uncompress tar.gz files}
\usage{
join_uncompress_targz(split_targz_files)
}
\arguments{
\item{split_targz_files}{character vector containing the full path values of the tar.gz files to be joint together.}
}
\value{
None
}
\description{
\code{join_uncompress_targz} joins multiple split tar.gz files to produce a single uncompressed file
}
\details{
\code{join_uncompress_targz} makes an internal system call to the program \code{tar} (\code{bsdtar}/GNU \code{tar}).
}
