% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoding.R
\name{one_hot_encoding}
\alias{one_hot_encoding}
\title{oneHotEncoding}
\usage{
one_hot_encoding(variant_descriptors)
}
\arguments{
\item{variant_descriptors}{Matrix containing categorical variables}
}
\value{
Matrix with values one-hot encoded
}
\description{
oneHotEncoding
}
