% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_variants.R
\name{run_variant_calling}
\alias{run_variant_calling}
\title{Run variant calling}
\usage{
run_variant_calling(bamfile)
}
\arguments{
\item{bamfile}{The filename of the BAM file}
}
\value{
Nothing
}
\description{
Runs a variant calling using Mutect2 by a system call... Annotates ...
}
\examples{
run_variant_calling(system.file("extdata", "sample.bam", package = "ideafix"))

}
