% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_deaminations.R
\name{annotate_deaminations}
\alias{annotate_deaminations}
\title{Annotate deaminations to file}
\usage{
annotate_deaminations(
  classification,
  format = "tsv",
  outfolder = ".",
  outname = "ideafix_labels",
  vcf_filename = NULL
)
}
\arguments{
\item{classification}{tibble containing the classification generated by \code{\link{classify_variants}}. It is the object to be written.}

\item{format}{character string indicating the output file format. Can be "tsv" or "vcf". Defaults to "tsv".}

\item{outfolder}{character string naming the folder to write the file to. Defaults to current working directory (\code{getwd}).}

\item{outname}{character string naming the output filename. Defaults to "ideafix_labels.txt" or "ideafix_labels.vcf", depending on \code{format} value .}

\item{vcf_filename}{optional character string, only needed if \code{format} equals to "vcf". It indicates the path to the vcf file ideafix has been run over.}
}
\value{
None
}
\description{
\code{annotate_deaminations} exports ideafix variant classification results to a text file. It either writes them to a new tsv file or appends them to the input vcf file provided to ideafix.
}
\details{

}
