% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_variants.R
\name{check_algorithm}
\alias{check_algorithm}
\title{Check classification algorithm name}
\usage{
check_algorithm(algorithm)
}
\arguments{
\item{algorithm}{Character string naming the algorithm to use to classify the variants.}
}
\value{
None
}
\description{
\code{check_algorithm} checks if a valid \code{algorithm} value has been provided.
}
\details{
Valid \code{algorithm} values are "RF" and "XGBoost". Valid value check is case-insensitive. For instance, "RF", "rf", "Rf" and "rF" are all valid \code{algorithm} argument values.

If an invalid \code{algorithm} value is provided, an error-message is displayed and the process is stopped.
}
