% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_variants.R
\name{check_descriptors}
\alias{check_descriptors}
\title{Check available descriptors}
\usage{
check_descriptors(variant_descriptors, algorithm)
}
\arguments{
\item{variant_descriptors}{tibble containing the variants to be classified together with their values for a series of descriptors obtained by \code{\link{get_descriptors}}.}

\item{algorithm}{character string naming the algorithm to use to classify the variants. Can be "RF" or "XGBoost".}
}
\value{
None
}
\description{
\code{check_descriptors} checks which variant descriptors are present in the input object \code{variant_descriptors} and decides if these are sufficient to be used with the model provided in \code{algorithm}.
}
\details{
\code{check_descriptors} checks if the descriptors present in \code{variant_descriptors} are sufficient to run the model in \code{algorithm} on them. If \code{variant_descriptors} object contains fewer descriptors than those included in the models and \code{algorithm} is "XGBoost", it throws an error message warning about the impossibility of executing the classification using that algorithm, and if the classification has been called, it stops. If \code{algorithm} equals to "RF" instead, a warning message alerts about this and the absence of performance guarantees, but the classification is executed if called.
}
