% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{extract_fa}
\alias{extract_fa}
\title{Extract descriptors from fasta file}
\usage{
extract_fa(mutation_ids, fasta_filename, k = 2)
}
\arguments{
\item{mutation_ids}{Vector that contains the mutation ids of the variants to be subjected to filtering. Mutation id format is chr:pos.}

\item{fasta_filename}{Reference genome data was aligned to.}

\item{k}{Number of bases to the right and to the left of the variant locus to get the genomic sequence from.}
}
\value{
Tibble containing the base composition of k positions before and k positions after the variant locus. Columns are mutation_id, current, bases, before.2, before.1, after.1, after.2, before and after.
}
\description{
\code{extract_fa} extracts the descriptors of the data that are derived from a reference genome fasta file
}
