% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{extract_vcf}
\alias{extract_vcf}
\title{Extract descriptors from vcf file}
\usage{
extract_vcf(vcf_filename, mutation_ids, samplename)
}
\arguments{
\item{vcf_filename}{VCF filename to extract the sample name from}

\item{mutation_ids}{Vector that contains the mutation ids of the variants to be subjected to filtering. Mutation id format is chr:pos.}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}
}
\value{
A tibble containing the descriptors of the variants that are extracted from vcf files.
}
\description{
Extract model descriptors that are found in vcf files
}
