% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_base_qual}
\alias{get_base_qual}
\title{Get base quality}
\usage{
get_base_qual(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}

\item{...}{Not to be checked}
}
\value{
List containing two vectors: median base quality of the alternate allele and ratio between median base quality of the alternate allele and median base quality of the reference allele
}
\description{
Get median Phred quality of the bases that support the alternate allele.
}
