% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_depth}
\alias{get_depth}
\title{Retrieve read depth}
\usage{
get_depth(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{character string naming the path to the input vcf, i.e. the vcf file containing the variants to classify. This file must have been generated with Mutect2, either in tumor only or tumor/normal mode with strand bias annotation enabled.}

\item{samplename}{character string naming the sample in \code{vcf_filename}. This must match the name given to the sample when running Mutect2. It can be obtained by calling \code{get_samplename} on \code{vcf_filename}.}

\item{...}{additional arguments not to be checked.}
}
\value{
Numeric vector with the read depth at each C:G > T:A variant locus present in \code{vcf_filename}, ordered by appearance in the file.
}
\description{
This function retrieves the read depth at each C:G > T:A variant locus present in \code{vcf_filename}. This data is retrieved from \code{vcf_filename}.
}
