% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_descriptors}
\alias{get_descriptors}
\title{Get descriptors}
\usage{
get_descriptors(vcf_filename, fasta_filename)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{fasta_filename}{Reference genome data was aligned to.}
}
\value{
A tibble containing the descriptors of the C:G > T:A variants needed for running the filtering.
}
\description{
Get all the C:G > T:A variant descriptors needed for running the filtering.
}
\details{
Get all the C:G > T:A variant descriptors needed for running the filtering. The descriptors are obtained from the variant vcf file and reference genome fasta file. The function can only run on a single vcf file.
}
