% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_map_qual}
\alias{get_map_qual}
\title{Get mapping quality}
\usage{
get_map_qual(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}

\item{...}{Not to be checked}
}
\value{
Vector containing the mapping quality of each variant, ordered in order of appearance in the vcf file
}
\description{
Get median Phred mapping quality (MAPQ) of the bases that support the alternate allele
}
