% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_mut_id}
\alias{get_mut_id}
\title{Get mutation id}
\usage{
get_mut_id(vcf_filename)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}
}
\value{
Vector containing the mutation id of the variants to be subjected to filtering. Mutation id format is samplename:chr:pos.
}
\description{
Parses a vcf file and creates a vector of the mutation ids of the variants to be subjected to filtering
}
