% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_pos_from_end}
\alias{get_pos_from_end}
\title{Get distance from end of read}
\usage{
get_pos_from_end(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}

\item{...}{Not to be checked}
}
\value{
Vector containing the median distance from end of read of each variant, ordered in order of appearance in the vcf file
}
\description{
Get median value of the position of the variant (starting from the end of the read) in reads that support the alternate allele.
}
