% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_ref_allele}
\alias{get_ref_allele}
\title{Retrieve reference allele}
\usage{
get_ref_allele(vcf_filename, ...)
}
\arguments{
\item{vcf_filename}{character string naming the path to the input vcf, i.e. the vcf file containing the variants to classify. This file must have been generated with Mutect2, either in tumor only or tumor/normal mode.}

\item{...}{additional arguments not to be checked.}
}
\value{
Character vector with the reference allele of each C:G > T:A variant present in \code{vcf_filename}, ordered by appearance in the file.
}
\description{
This function retrieves the reference allele of each C:G > T:A variant present in \code{vcf_filename}. This data is retrieved from \code{vcf_filename}.
}
