% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_RPOB}
\alias{get_RPOB}
\title{Get read pair orientation bias}
\usage{
get_RPOB(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}

\item{...}{Not to be checked}
}
\value{
List containing two vectors: for each variant, FdeamC and SOB
}
\description{
Get two read pair orientation bias (RPOB) metrics: FdeamC and strand orientation bias
}
