% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_samplename}
\alias{get_samplename}
\title{Get sample name from vcf file}
\usage{
get_samplename(vcf_filename)
}
\arguments{
\item{vcf_filename}{VCF filename to extract the sample name from}
}
\value{
Name given to the sample the vcf filename contains variants from
}
\description{
It extracts from the VCF filename the name given to the sample in which variants were called
}
\details{
It corresponds to the tumor field in the vcf filename. In tumor-only variant calling, that is the only sample. In tumor-normal variant calling, it corresponds to the tumor sample
}
