% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{get_SB}
\alias{get_SB}
\title{Get strand bias}
\usage{
get_SB(vcf_filename, samplename, ...)
}
\arguments{
\item{vcf_filename}{Variant call vcf filename}

\item{samplename}{Name given to the sample the vcf filename contains variants from, ordered in order of appearance in the vcf file}

\item{...}{Not to be checked}
}
\value{
List containing two vectors: for each variant, strand bias defined by GATK and strand bias defined by Guo et al.
}
\description{
Get two strand bias (SB) metrics: SB defined by GATK and SB defined by Guo et al.
}
