% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{join_uncompress_targz}
\alias{join_uncompress_targz}
\title{Join split tar files and uncompress them}
\usage{
join_uncompress_targz(split_targz_files)
}
\arguments{
\item{split_targz_files}{Vector containing the full paths of the parts in which a tar file has been split}
}
\value{
None
}
\description{
Join split tar files and uncompress them
}
