% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptors.R
\name{write_region_file}
\alias{write_region_file}
\title{Write temporary locus region file}
\usage{
write_region_file(mutation_ids, k, tmp_filename)
}
\arguments{
\item{mutation_ids}{Vector containing the mutation id of the variants to be subjected to filtering. Mutation id format is samplename:chr:pos.}

\item{k}{Number of bases to the right and to the left of the variant locus to get the genomic sequence from.}

\item{tmp_filename}{Path of the temporary file.}
}
\value{
None
}
\description{
Creates the file needed by samtools faidx to get the surrounding bases of a locus
}
