% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_deaminations.R
\name{write_vcf}
\alias{write_vcf}
\title{Append filtering results to input vcf file}
\usage{
write_vcf(vcf_filename, predictions, outname)
}
\arguments{
\item{vcf_filename}{input vcf}

\item{predictions}{predictions generated by filter_variants}
}
\value{

}
\description{
Append filtering results to input vcf file
}
